requirements.txt
# Core data processing
beautifulsoup4>=4.10.0
requests>=2.27.0
selenium>=4.1.0
pandas>=1.3.0
numpy>=1.20.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# Scheduling and automation
schedule>=1.1.0
python-dotenv>=0.19.0

# Web scraping utilities
webdriver-manager>=3.5.0

# Progress bars and utilities
tqdm>=4.62.0

# Jupyter notebooks
jupyter>=1.0.0
ipython>=7.30.0
